+++ src/app/features/sections/privacy-policy/privacy-policy.component.ts
@@ -1,0 +1,12 @@
+import { Component } from '@angular/core';
+import { TranslatePipe} from '@ngx-translate/core';
+
+@Component({
+  selector: 'app-privacy-policy',
+  imports: [TranslatePipe],
+  templateUrl: './privacy-policy.component.html',
+  styleUrl: './privacy-policy.component.css'
+})
+export class PrivacyPolicyComponent {
+
+}